+++ src/components/charge/index.js
@@ -1,0 +1,9 @@
+/**
+ * @fileoverview Index des composants Charge Technique
+ * @module components/charge
+ */
+
+export { ChargesSynthesisCards } from './ChargesSynthesisCards';
+export { CauseBreakdownSection } from './CauseBreakdownSection';
+export { EquipementClassTable } from './EquipementClassTable';
+export { DecisionGuide } from './DecisionGuide';